requirements.txt
pandas==1.0.1
numpy==1.18.1
requests==2.23.0
python-whois==0.7.3
fuzzywuzzy==0.18.0